+++ crates/swc/tests/tsc.rs
@@ -14,9 +14,7 @@ use regex::Regex;
 use serde::de::DeserializeOwned;
 use serde_json::from_str;
 use swc::{
-    config::{
-        Config, IsModule, JsMinifyOptions, JscConfig, ModuleConfig, Options, TransformConfig,
-    },
+    config::{Config, JsMinifyOptions, JscConfig, ModuleConfig, Options, TransformConfig},
     try_with_handler, Compiler,
 };
 use swc_common::{
@@ -393,7 +391,6 @@ fn matrix(input: &Path) -> Vec<TestUnitData> {
                             ..Default::default()
                         },
                         module: Some(module.into()),
-                        is_module: IsModule::Bool(true),
                         ..Default::default()
                     },
                     ..Default::default()