+++ crates/swc_ecma_minifier/tests/terser/compress/harmony/expansion/output.mangleOnly.js
@@ -1,3 +1,3 @@
-function o(o, ...n) {
-    console.log(o);
+function o(n, ...c) {
+    console.log(n);
 }