+++ crates/swc_ecma_transforms_base/tests/resolver/issues/2854/1/output.js
@@ -1,5 +1,5 @@
 export function App__1() {
-    return <Form__1 />;
+    return <Form__1/>;
 }
 export function Form__1({ onChange__2 =function() {}  }) {
     return <input onChange__0={function() {